third_party/blink/web_tests/external/wpt/css/css-display/display-math-on-pseudo-elements-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>display: math and inline-math on pseudo elements compute to flow</title>
<link rel="help" href="https://drafts.csswg.org/css-display/#the-display-properties">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  [data-expected="inline"]::before { content: 'math'; display: math; }
  [data-expected="inline"]::after { content: 'math'; display: math; }
  [data-expected="block"]::before { content: 'math'; display: block math; }
  [data-expected="block"]::after { content: 'math'; display: block math; }
</style>
<div>
  <div data-expected="block"></div>
  <math data-expected="block"></math>
  <math><mrow data-expected="block"></mrow></math>
  <div data-expected="inline"></div>
  <math data-expected="inline"></math>
  <math><mrow data-expected="inline"></mrow></math>
</div>
<script>
  Array.from(document.querySelectorAll('[data-expected]')).forEach(element => {
      test(function() {
          assert_equals(window.getComputedStyle(element, "::before").display,
                        element.getAttribute('data-expected'));
          assert_equals(window.getComputedStyle(element, "::after").display,
                        element.getAttribute('data-expected'));
      }, `computed display on ::before and ::after for <${element.tagName} class="${element.getAttribute('class')}">`);
  });
</script>
